Fixing merging issues with get_data_bytes
