+++ src/qrl/core/txs/LatticePublicKey.py
@@ -23,13 +23,11 @@ class LatticePublicKey(Transaction):
     def dilithium_pk(self):
         return self._data.latticePK.dilithium_pk
 
-    def get_hashable_bytes(self):
-        return sha256(
-            self.master_addr +
-            self.fee.to_bytes(8, byteorder='big', signed=False) +
-            self.kyber_pk +
-            self.dilithium_pk
-        )
+    def get_data_bytes(self) -> bytes:
+        return self.master_addr + \
+               self.fee.to_bytes(8, byteorder='big', signed=False) + \
+               self.kyber_pk + \
+               self.dilithium_pk
 
     @staticmethod
     def create(fee, kyber_pk, dilithium_pk, xmss_pk, master_addr: bytes = None):
