+++ src/qrl/core/txs/Transaction.py
@@ -134,20 +134,27 @@ class Transaction(object, metaclass=ABCMeta):
 
     def update_txhash(self):
         self._data.transaction_hash = sha256(
-            self.get_hashable_bytes() +
+            self.get_data_hash() +
             self.signature +
             self.PK
         )
 
-    def get_hashable_bytes(self) -> bytes:
+    def get_data_bytes(self) -> bytes:
         """
-        This method returns the hashes of the transaction data.
+        This method returns the essential bytes that represent the transaction and will be later signed
         :return:
         """
         raise NotImplementedError
 
-    def sign(self, xmss):
-        self._data.signature = xmss.sign(self.get_hashable_bytes())
+    def get_data_hash(self) -> bytes:
+        """
+        This method returns the hashes of the transaction data.
+        :return:
+        """
+        return sha256(self.get_data_bytes())
+
+    def sign(self, object_with_sign_method):
+        self._data.signature = object_with_sign_method.sign(self.get_data_hash())
         self.update_txhash()
 
     @abstractmethod
@@ -211,7 +218,7 @@ class Transaction(object, metaclass=ABCMeta):
 
         return True
 
-    def validate(self) -> bool:
+    def validate(self, verify_signature=True) -> bool:
         """
         This method calls validate_or_raise, logs any failure and returns True or False accordingly
         The main purpose is to avoid exceptions and accommodate legacy code
@@ -219,7 +226,7 @@ class Transaction(object, metaclass=ABCMeta):
         :rtype: bool
         """
         try:
-            self.validate_or_raise()
+            self.validate_or_raise(verify_signature)
         except ValueError as e:
             logger.info('[%s] failed validate_tx', bin2hstr(self.txhash))
             logger.warning(str(e))
@@ -239,7 +246,7 @@ class Transaction(object, metaclass=ABCMeta):
     def _get_master_address(self):
         return self.addr_from
 
-    def validate_or_raise(self) -> bool:
+    def validate_or_raise(self, verify_signature=True) -> bool:
         """
         This method will validate a transaction and raise exception if problems are found
         :return: True if the exception is valid, exceptions otherwise
@@ -250,10 +257,11 @@ class Transaction(object, metaclass=ABCMeta):
 
         self._coinbase_filter()
 
-        if not XmssFast.verify(self.get_hashable_bytes(),
-                               self.signature,
-                               self.PK):
+        if verify_signature and not XmssFast.verify(self.get_data_hash(),
+                                                    self.signature,
+                                                    self.PK):
             raise ValueError("Invalid xmss signature")
+
         return True
 
     def validate_slave(self, addr_from_state: AddressState, addr_from_pk_state: AddressState):
