+++ src/qrl/core/txs/TransferTokenTransaction.py
@@ -35,17 +35,17 @@ class TransferTokenTransaction(Transaction):
     def amounts(self):
         return self._data.transfer_token.amounts
 
-    def get_hashable_bytes(self):
-        tmptxhash = (self.master_addr +
-                     self.fee.to_bytes(8, byteorder='big', signed=False) +
-                     self.token_txhash)
+    def get_data_bytes(self):
+        data_bytes = (self.master_addr +
+                      self.fee.to_bytes(8, byteorder='big', signed=False) +
+                      self.token_txhash)
 
         for index in range(0, len(self.addrs_to)):
-            tmptxhash = (tmptxhash +
-                         self.addrs_to[index] +
-                         self.amounts[index].to_bytes(8, byteorder='big', signed=False))
+            data_bytes = (data_bytes +
+                          self.addrs_to[index] +
+                          self.amounts[index].to_bytes(8, byteorder='big', signed=False))
 
-        return sha256(tmptxhash)
+        return data_bytes
 
     @staticmethod
     def create(token_txhash: bytes,
